back-end/requirements.txt
# requirements
Django
python-decouple
djangorestframework
psycopg2-binary
django-cors-headers
django-filter
